+++ EvilBot/TwitchInfo.Designer.cs
@@ -19,7 +19,7 @@ namespace EvilBot {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "15.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "4.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class TwitchInfo {
@@ -61,7 +61,7 @@ namespace EvilBot {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to 48efkpjcvpg6v69vhk8v866lktsyce.
+        ///   Looks up a localized string similar to hnrw2r6sviculzzduq8rguap5ycate.
         /// </summary>
         internal static string BotToken {
             get {